Programa Terminado

+++ src/ejercicioCategoria/Categoria.java
@@ -2,6 +2,6 @@ package ejercicioCategoria;
 
 public abstract class Categoria {
 
-    public abstract int sueldoPromedio(Usuario user);
+    public abstract double sueldoPromedio(Usuario user);
 
 }
